cleanup example logic
